Add simple KSQL query to test transformations.
This works for the selected topic, given the appropriate schema definition within
the 'CREATE STREAM' command, however it appears to be returning only every fourth
record from the topic.  Also, some values are out of order -- need to check the
source data to see if this is a Kafka issue.